refactor(graph): extract the excerpt indexer + shared index helpers (#1624)
First of the per-format `indexers.ts` splits (arch D1, "one format per PR").

- New `graph/index-helpers.ts` (leaf): the two low-level helpers every format
  needs — `fileMtimeIso` + `injectPrefixes` — moved out of indexers.ts so the
  format modules can import them without depending back on the big file. Depends
  only on `./state`, so the split stays acyclic.
- New `graph/indexers/excerpt.ts`: the excerpt cluster — indexExcerpt,
  removeExcerpt, excerptIdsForSource, parseExcerptIdFromPath — plus its
  `walkAndIndexExcerpts` walker (moved with it so indexAllNotes imports it rather
  than the other way round, avoiding a cycle).
- indexers.ts imports the two helpers + the walker; the graph facade re-exports
  the excerpt functions from the new module. External callers (via the facade)
  are unchanged. indexers.ts: 1669 → 1559 lines.

Pure relocation — the full graph suite (440 tests) is green, lint clean.
(Removing the excerpt code also made the flaky require-await stop firing on
indexNote, so the misplaced #1624 disable directive is dropped.)

Follow-ups: extract the source, tables, and note-file-type indexers the same way.
